fixed bug where program would hang on exit. the bug was that i had declared a default constructor in the sub manager which you are not supposed to do if you have a member reference variable so it lead to undefined behavior

+++ GUI/include/Widget.hpp
@@ -17,7 +17,7 @@ class Widget {
       window_id = randomID(mt); 
     }
 
-    int getWindowID() const { return window_id; }
+    int64_t getWindowID() const { return window_id; }
     bool isOpen() { return show; }
 
     virtual ~Widget() = default;
